media_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
boto3==1.34.34
python-dotenv==1.0.1
Pillow==10.2.0
aiokafka==0.12.0